+++ CHANGELOG.rst
@@ -2,11 +2,13 @@
 Changelog
 =========
 
-0.2.2 (WIP)
------------
+0.3.0 (2020-12-15)
+------------------
 
 * Add (private) scripts (make venv, run checks).
 * Update usage docs.
+* Extend / rewrite discord client methods.
+* Reuse existing tqdm :class:`transformers.trainer_callback.ProgressCallback` for progress tracking
 
 0.2.1 (2020-12-15)
 ------------------